RAG_TEXT-TO-SQL/requirements.txt
streamlit
python-dotenv
pymupdf
Pillow
google-generativeai
langchain
sqlalchemy
faiss-cpu
langchain-community
sentence-transformers
torch
torch
torchvision
torchaudio